// frontend/src/lib/api/generated/schemas/listLeadsFrontendApiTenantsTenantIdLeadsGetParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Pulsar Platform API
 * Multi-tenant Telegram lead parsing platform
 * OpenAPI spec version: 1.0.0
 */

export type ListLeadsFrontendApiTenantsTenantIdLeadsGetParams = {
page?: number;
page_size?: number;
search?: string | null;
status?: string | null;
tag?: string | null;
source?: string | null;
period?: string | null;
sort_by?: string;
sort_dir?: string;
};