// frontend/src/lib/api/generated/schemas/bulkUpdateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Pulsar Platform API
 * Multi-tenant Telegram lead parsing platform
 * OpenAPI spec version: 1.0.0
 */
import type { BulkActionPayload } from './bulkActionPayload';
import type { BulkUpdateInputFilters } from './bulkUpdateInputFilters';

export interface BulkUpdateInput {
  ids?: number[] | null;
  applyToFiltered?: boolean;
  filters?: BulkUpdateInputFilters;
  action: BulkActionPayload;
}